scf_loop: modularize and clean up
